C++ API: (CLEC) - move functions L_cc2(), L_cc_stream(), L_cc() and                   L_solve() as CLEC constructors                 - rename function L_cc1() as generate_lec_expression()                 - rename source file l_cc1.c as l_compile.c                 - delete source files l_cc2.c and l_err.c
part of issue #1029

+++ api/lec/l_exec_mtfn.h
@@ -165,7 +165,7 @@ public:
         pos_buffer += 2 + (1 + nv_args) * sizeof(short);
     }
 
-    int get_length() const override 
+    short get_length() const override 
     {
         // 3 bytes for type, nb_args and nv_args + sizeof(short) for len_args + 
         // len_args for the arguments
